train/requirements.txt
# BoxWorld TransUNet 训练依赖

# PyTorch (根据你的 CUDA 版本选择)
# 访问 https://pytorch.org/get-started/locally/ 获取安装命令
torch>=2.0.0
torchvision>=0.15.0

# 图像处理
Pillow>=9.0.0
opencv-python>=4.5.0

# 数据增强
albumentations>=1.3.0

# 科学计算
numpy>=1.21.0
scipy>=1.7.0

# 进度条
tqdm>=4.60.0

# 可视化
matplotlib>=3.5.0

# 可选: 实验跟踪
# wandb>=0.15.0

# 可选: TensorBoard
# tensorboard>=2.10.0

# 开发工具
# black>=22.0.0
# isort>=5.10.0
